Handle unselected spawn points for players
Players who do not select a spawn point are now assigned a random spawn from their party members or from the available spawnpoints map. If no spawn point is assigned, they are randomly teleported in the prepare state. Also, the HideSpawnSelection RPC is resent in case some players missed it.